requirements.txt
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy == 1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==24.2
pandas==2.0.3
patsy==1.0.1
pillow==10.4.0
pyparsing==3.1.4
python == 3.10
pytz==2024.2
scikit-learn==1.3.2
scikit-network==0.32.1
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
statsmodels==0.14.1
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
tslearn==0.6.3
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.20.2
